requirements.txt
# Third-party packages with pinned versions
python-dotenv~=1.1.1
pyngrok~=7.3.0
chromadb~=1.1.0
tiktoken~=0.11.0
requests~=2.32.5
redis~=5.0.0
fastapi~=0.116.2
python-telegram-bot
pdfplumber~=0.10.0
celery~=5.3.0
openai~=1.0.0
pytesseract~=0.3.10
pdf2image~=1.16.0
python-telegram-bot
ijson
